+++ examples/data-sources/coolify_version/data-source.tf
@@ -1,0 +1,5 @@
+data "coolify_version" "current" {}
+
+output "coolify_version" {
+  value = data.coolify_version.current.version
+}