+++ internal/provider/provider.go
@@ -22,6 +22,7 @@ import (
 	"github.com/SebTardif/terraform-provider-coolify/internal/service/server"
 	"github.com/SebTardif/terraform-provider-coolify/internal/service/service"
 	"github.com/SebTardif/terraform-provider-coolify/internal/service/team"
+	"github.com/SebTardif/terraform-provider-coolify/internal/service/version"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/provider/schema"
@@ -75,6 +76,9 @@ func (p *coolifyProvider) Configure(ctx context.Context, req provider.ConfigureR
 		return
 	}
 	c := client.New(endpoint, token)
+	if p.version != "" {
+		c.UserAgent = "terraform-provider-coolify/" + p.version
+	}
 	resp.DataSourceData = c
 	resp.ResourceData = c
 }
@@ -82,5 +86,5 @@ func (p *coolifyProvider) Resources(_ context.Context) []func() resource.Resourc
 	return []func() resource.Resource{application.NewResource, application.NewDockerResource, application.NewDockerComposeResource, application.NewPrivateGitResource, backup.NewResource, deployment.NewResource, environmentvariable.NewResource, postgresql.NewResource, mysql.NewResource, mariadb.NewResource, redis.NewResource, mongodb.NewResource, clickhouse.NewResource, keydb.NewResource, dragonfly.NewResource, privatekey.NewResource, project.NewResource, server.NewResource, service.NewResource}
 }
 func (p *coolifyProvider) DataSources(_ context.Context) []func() datasource.DataSource {
-	return []func() datasource.DataSource{application.NewDataSource, application.NewListDataSource, database.NewListDataSource, database.NewDataSource, project.NewDataSource, project.NewListDataSource, server.NewDataSource, server.NewListDataSource, server.NewResourcesDataSource, server.NewDomainsDataSource, service.NewListDataSource, service.NewDataSource, privatekey.NewDataSource, privatekey.NewListDataSource, team.NewDataSource}
+	return []func() datasource.DataSource{application.NewDataSource, application.NewListDataSource, database.NewListDataSource, database.NewDataSource, project.NewDataSource, project.NewListDataSource, server.NewDataSource, server.NewListDataSource, server.NewResourcesDataSource, server.NewDomainsDataSource, service.NewListDataSource, service.NewDataSource, privatekey.NewDataSource, privatekey.NewListDataSource, team.NewDataSource, version.NewDataSource}
 }